Scene Change
Trying to add in scene changing functionality from Main Menu to just 360 Video with controls and a Home button to go back to the Main Menu.
